requirements.txt
baidu-aip==2.2.13.0
certifi==2019.3.9
chardet==3.0.4
idna==2.8
requests==2.21.0
urllib3==1.24.1
